Revert "Trust merge"
This reverts commit 2f43ce0c2454f19cd2988294fd0786285f450756.

+++ src/map/utils/battlefieldutils.h
@@ -34,7 +34,7 @@ namespace battlefieldutils
 {
     void getLosePosition(CBattlefield* battlefield, int(&pPosition)[4]);		// returns x y z rot in that order
     void getWinPosition(CBattlefield* battlefield, int(&pPosition)[4]);		// returns x y z rot in that order
-    void getStartPosition(uint16 zoneid, float(&pPosition)[4]);			// returns lobby position
+    void getStartPosition(uint16 zoneid, int(&pPosition)[4]);			// returns lobby position
 };
 
 #endif
